# GH-3789 RDFS Reasoner Challenge fixes