--- a/core/sail/base/src/main/java/org/eclipse/rdf4j/sail/base/SailSourceConnection.java
+++ b/core/sail/base/src/main/java/org/eclipse/rdf4j/sail/base/SailSourceConnection.java
@@ -669,31 +669,40 @@ public abstract class SailSourceConnection extends AbstractNotifyingSailConnecti
 				explicitOnlyDataset = branch(IncludeInferred.explicitOnly).dataset(level);
 			}
 			boolean modified = false;
-			if (contexts.length == 0) {
+			if (contexts.length == 0 || contexts.length == 1 && contexts[0] == null) {
 				if (!hasStatement(explicitOnlyDataset, subj, pred, obj, NULL_CTX)) {
 					// only add inferred statements that aren't already explicit
-					if (!hasStatement(inferredOnlyDataset, subj, pred, obj, NULL_CTX)) {
+					boolean notHasStatement = !hasStatement(inferredOnlyDataset, subj, pred, obj, NULL_CTX);
+					inferredOnlySink.approve(subj, pred, obj, null);
+					if (notHasStatement) {
 						// only report inferred statements that don't already
 						// exist
 						addStatementInternal(subj, pred, obj, contexts);
 						notifyStatementAdded(vf.createStatement(subj, pred, obj));
 						modified = true;
 					}
-					inferredOnlySink.approve(subj, pred, obj, null);
 				}
 			} else {
 				for (Resource ctx : contexts) {
-					if (!hasStatement(explicitOnlyDataset, subj, pred, obj, ctx)) {
+					Resource[] contextsToCheck;
+					if (contexts.length == 1) {
+						contextsToCheck = contexts;
+					} else {
+						contextsToCheck = new Resource[] { ctx };
+					}
+
+					if (!hasStatement(explicitOnlyDataset, subj, pred, obj, contextsToCheck)) {
 						// only add inferred statements that aren't already
 						// explicit
-						if (!hasStatement(inferredOnlyDataset, subj, pred, obj, ctx)) {
+						boolean notHasStatement = !hasStatement(inferredOnlyDataset, subj, pred, obj, contextsToCheck);
+						inferredOnlySink.approve(subj, pred, obj, ctx);
+						if (notHasStatement) {
 							// only report inferred statements that don't
 							// already exist
 							addStatementInternal(subj, pred, obj, ctx);
 							notifyStatementAdded(vf.createStatement(subj, pred, obj, ctx));
 							modified = true;
 						}
-						inferredOnlySink.approve(subj, pred, obj, ctx);
 					}
 				}
 			}