@hmottestad 2022-04-11
### Before
```

Benchmark                                                                                        (param)  Mode  Cnt   Score   Error  Units
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                            moreRdfs::12180  avgt    3  19.329 ± 3.140  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                            longChain::5803  avgt    3  17.389 ± 5.664  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                                medium::544  avgt    3   1.279 ± 0.048  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencer                                simple::152  avgt    3   0.791 ± 0.025  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions        moreRdfs::12180  avgt    3  13.677 ± 1.032  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions        longChain::5803  avgt    3  16.933 ± 2.610  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions            medium::544  avgt    3   1.127 ± 0.041  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactions            simple::152  avgt    3   0.782 ± 0.026  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema  moreRdfs::12180  avgt    3  12.675 ± 0.802  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema  longChain::5803  avgt    3  13.176 ± 0.493  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema      medium::544  avgt    3   0.618 ± 0.041  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerMultipleTransactionsSchema      simple::152  avgt    3   0.292 ± 0.010  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                      moreRdfs::12180  avgt    3  13.574 ± 0.567  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                      longChain::5803  avgt    3  12.763 ± 0.394  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                          medium::544  avgt    3   0.620 ± 0.013  ms/op
ReasoningBenchmark.forwardChainingSchemaCachingRDFSInferencerSchema                          simple::152  avgt    3   0.294 ± 0.008  ms/op
ReasoningBenchmark.noReasoning                                                           moreRdfs::12180  avgt    3   4.173 ± 0.871  ms/op
ReasoningBenchmark.noReasoning                                                           longChain::5803  avgt    3   1.821 ± 0.058  ms/op
ReasoningBenchmark.noReasoning                                                               medium::544  avgt    3   0.291 ± 0.008  ms/op
ReasoningBenchmark.noReasoning                                                               simple::152  avgt    3   0.124 ± 0.015  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                       moreRdfs::12180  avgt    3   4.044 ± 0.383  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                       longChain::5803  avgt    3   1.819 ± 0.376  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                           medium::544  avgt    3   0.292 ± 0.040  ms/op
ReasoningBenchmark.noReasoningMultipleTransactions                                           simple::152  avgt    3   0.125 ± 0.008  ms/op
```

### After
```

```